chore: move changes to proxy
