+++ models/classes/search/GenerisSearchBridge.php
@@ -20,17 +20,43 @@
 
 namespace oat\tao\model\search;
 
+use oat\generis\model\OntologyAwareTrait;
 use oat\oatbox\service\ConfigurableService;
 
 class GenerisSearchBridge extends ConfigurableService implements SearchBridgeInterface
 {
+    use OntologyAwareTrait;
+
     public function search(SearchQuery $query): ResultSet
     {
+        if ($this->isUriSearch($query->getTerm())) {
+            $resource = $this->getResource($query->getTerm());
+            $class = $this->getClass($query->getParentClass());
+
+            if ($resource->exists() && $resource->isInstanceOf($class)) {
+                return new ResultSet(
+                    [
+                        [
+                            'id' => $resource->getUri(),
+                            'label' => $resource->getLabel(),
+                        ],
+                    ],
+                    1
+                );
+            }
+        }
+
         return $this->getServiceLocator()->get(Search::SERVICE_ID)->query(
             $query->getTerm(),
             $query->getParentClass(),
             $query->getStartRow(),
             $query->getRows()
         );
+    }
+
+    private function isUriSearch(string $queryString): bool
+    {
+        return strpos($queryString, LOCAL_NAMESPACE) === 0
+            || filter_var($queryString, FILTER_VALIDATE_URL) !== false;
     }
 }
